+++ CMakeLists.txt
@@ -172,4 +172,6 @@ add_subdirectory(apps/test-moisture)
 add_subdirectory(apps/test-pid)
 add_subdirectory(apps/test-pwm)
 add_subdirectory(apps/test-pwmin)
-add_subdirectory(apps/tutorial_servo_control)
+add_subdirectory(apps/tutorial-servo-pot-control)
+add_subdirectory(apps/tutorial-servo-can-control)
+
