+++ apps/tutorial-servo-pot-control/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_executable(tutorial-servo-pot-control.${TARGET}-board.elf)
+target_sources(tutorial-servo-pot-control.${TARGET}-board.elf PRIVATEsrc/main.cpp)
+target_set_firmware_properties(tutorial-servo-pot-control.${TARGET}-board.elf)
